+++ Interviews.Services/Models/Applicants.cs
@@ -6,7 +6,7 @@ namespace Interviews.Services.Models
 {
     public class Applicants
     {
-        public long ApplicantId { get; set; }
+        public int ApplicantId { get; set; }
         public string Name { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
